9.117 protecting resources

+++ backend/routes/posts.js
@@ -114,9 +114,13 @@ router.put(
       imagePath: imagePath
     });
 
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      console.log(result);
-      res.status(200).json({message: 'Update successful!'});
+    Post.updateOne({ _id: req.params.id, creator:req.userData.userId }, post).then(result => {
+      if(result.nModified > 0) {
+        res.status(200).json({message: 'Update successful!'});
+      }
+      else {
+        res.status(401).json({message: 'Not authorized!'});
+      }
     });
 });
 
@@ -126,9 +130,13 @@ router.delete(
   (req, res, next) => {
     const id = req.params.id;
     console.log(id);
-    Post.deleteOne({ _id: id}).then(result => {
-      console.log(result);
-      res.status(200).json({message: 'Post deleted'})
+    Post.deleteOne({ _id: id, creator:req.userData.userId }).then(result => {
+      if(result.n > 0) {
+        res.status(200).json({message: 'Post deleted'})
+      }
+      else {
+        res.status(401).json({message: 'Not authorized!'});
+      }
     });
 });
